\name{calc_residuals}
\alias{calc_residuals}
\title{Calculate Unmixing Residuals}
\usage{calc_residuals(flow_frame, ref_matrix)}
\arguments{
\item{flow_frame}{flowFrame object with raw data}
\item{ref_matrix}{Reference matrix from build_ref_matrix}
}
\value{data.frame with abundances and RMSE_Score column}
