% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autospectral_interface.R
\name{extract_af_signatures}
\alias{extract_af_signatures}
\title{Internal helper to extract gated AF signatures}
\usage{
extract_af_signatures(af_dir, detector_names)
}
\description{
Internal helper to extract gated AF signatures
}
