% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing_matrix.R
\name{plot_unmixing_matrix}
\alias{plot_unmixing_matrix}
\title{Plot Unmixing Matrix}
\usage{
plot_unmixing_matrix(W, pd = NULL)
}
\arguments{
\item{W}{Unmixing matrix}

\item{pd}{Optional pData for descriptive labels}
}
\value{
ggplot object
}
\description{
Plot Unmixing Matrix
}
\examples{
M <- matrix(c(1, 0.2, 0.1, 1), nrow = 2, byrow = TRUE)
rownames(M) <- c("FITC", "PE")
colnames(M) <- c("B2-A", "YG1-A")
p <- plot_unmixing_matrix(M)
print(p)
}
