% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating_options.R
\name{get_sorted_detectors}
\alias{get_sorted_detectors}
\title{Get Sorted Detectors by Laser and Wavelength}
\usage{
get_sorted_detectors(pd)
}
\arguments{
\item{pd}{pData from flowFrame parameters}
}
\value{
A list with [[names]] (FL...) and [[labels]] (405nm - 420/10) sorted by laser.
}
\description{
Get Sorted Detectors by Laser and Wavelength
}
