\name{plot_marker_correlations}
\alias{plot_marker_correlations}
\title{Plot Marker vs RMSE Correlations}
\usage{plot_marker_correlations(data, markers = NULL)}
\arguments{
\item{data}{data.frame from calc_residuals}
\item{markers}{Vector of marker names to plot}
}
\value{ggplot object with facets}
